+++ src/components/input/TimeInput.js
@@ -4,10 +4,18 @@ import Icon from '../icon';
 import SingleLineInputBase from './SingleLineInputBase';
 import { IconTimerSmallOutline } from '@teamleader/ui-icons';
 
+const isValidTime = input => RegExp('([0-1][0-9]|[2][0-3]):([0-5][0-9])').test(input);
 class TimeInput extends PureComponent {
+  beforeMaskedValueChange = ({ value: newValue, selection }, { value: oldValue }) => {
+    if (!isValidTime(newValue)) {
+      return { value: oldValue, selection };
+    }
+    return { value: newValue, selection };
+  };
+
   render() {
     return (
-      <InputMask {...this.props} mask="99:99" maskChar="0">
+      <InputMask {...this.props} mask="99:99" maskChar="0" beforeMaskedValueChange={this.beforeMaskedValueChange}>
         {inputProps => (
           <SingleLineInputBase
             {...inputProps}